+++ test/test.dart
@@ -1,17 +1,11 @@
-import '../lib/src/util/log.dart';
-import '../lib/src/util/tiny_tester.dart';
-import '../lib/src/util/result.dart';
+import 'extension_test/date_extension_test.dart';
+import 'extension_test/list_extension_test.dart';
+import 'extension_test/number_extension_test.dart';
+import 'extension_test/string_extension_test.dart';
 
-import 'extension_test.dart';
-
-void main(List<String> args) {
-  extensionTest();
-  Log.runTest();
-  testResult();
-  Result.suppress();
-  Result.hasError(
-      value: 'Testing',
-      clientMessage: 'bruh',
-      devMessage: 'yoo',
-      errorType: ErrorTypes.other);
+void main() {
+  dateExtensionTest();
+  listExtensionTest();
+  numberExtensionTest();
+  stringExtensionTest();
 }